pyproject.toml
[project]
name = "gstats"
version = "0.1.0"
description = "display git stats in a streamlit app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dateutil>=2.9.0.post0",
    "streamlit>=1.44.1",
]
